+++ R/methods-epi_df.R
@@ -8,19 +8,22 @@
 #' use `attr(your_epi_df, "decay_to_tibble") <- FALSE` beforehand.
 #'
 #' @template x
-#' @param ... additional arguments to forward to `NextMethod()`
+#' @param ... forwarded to `as_tibble` for `data.frame`s
 #'
 #' @importFrom tibble as_tibble
 #' @export
 as_tibble.epi_df <- function(x, ...) {
-  # Decaying drops the class and metadata. `as_tibble.grouped_df` drops the
-  # grouping and should be called by `NextMethod()` in the current design.
-  # See #223 for discussion of alternatives.
+  # Note that some versions of `tsibble` overwrite `as_tibble.grouped_df`, which
+  # also impacts grouped `epi_df`s don't rely on `NextMethod()`. Destructure
+  # first instead.
+  destructured <- tibble::as_tibble(vctrs::vec_data(x), ...)
   if (attr(x, "decay_to_tibble") %||% TRUE) {
-    return(decay_epi_df(NextMethod()))
+    return(destructured)
+  } else {
+    # We specially requested via attr not to decay epi_df-ness but to drop any
+    # grouping.
+    reclass(destructured, attr(x, "metadata"))
   }
-  metadata <- attr(x, "metadata")
-  reclass(NextMethod(), metadata)
 }
 
 #' Convert to tsibble format