+++ man-roxygen/basic-slide-params.R
@@ -29,9 +29,6 @@
 #'   element of this vector serves as the reference time point for one sliding
 #'   window. If missing, then this will be set to all unique time values in the
 #'   underlying data table, by default.
-#' @param names_sep String specifying the separator to use in `tidyr::unnest()`
-#'   when `as_list_col = FALSE`. Default is "_". Using `NULL` drops the prefix
-#'   from `new_col_name` entirely.
 #' @param all_rows If `all_rows = TRUE`, then all rows of `x` will be kept in
 #'   the output even with `ref_time_values` provided, with some type of missing
 #'   value marker for the slide computation output column(s) for `time_value`s
@@ -42,5 +39,14 @@
 #'   the missing marker is a `NULL` entry in the list column; for certain
 #'   operations, you might want to replace these `NULL` entries with a different
 #'   `NA` marker.
-#' @return An `epi_df` object given by appending one or more new columns to
-#'  `x`, named according to the `new_col_name` argument.
+#' @param as_list_col `r lifecycle::badge("deprecated")` if you want a list
+#'   column as output, you can now just directly output a list from your slide
+#'   computations. Usually this just means wrapping your output in a length-1
+#'   list (outputting `list(result)` instead of `result`).
+#' @param names_sep `r lifecycle::badge("deprecated")` if you were specifying
+#'   `names_sep = NULL`, that's no longer needed. If you were using a non-NULL
+#'   value, you can either directly prefix your slide computation names, or
+#'   output a list and then later call `tidyr::unnest(slide_output,
+#'   <result_column_name>, names_sep = <names_sep>)`.
+#' @return An `epi_df` object given by appending one or more new columns to `x`,
+#'   named according to the `new_col_name` argument.